requirements.txt
phidata
python-dotenv
yfinance
packaging
duckduckgo-search
fastapi[standard]
uvicorn
groq
openai
python-multipart